+++ example/lib/json_place_holder/json_place_holder.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import './json_place_holder_view.dart';
 
 class JsonPlaceHolder extends StatefulWidget {
+  const JsonPlaceHolder({Key key}) : super(key: key);
+
   @override
   JsonPlaceHolderView createState() => JsonPlaceHolderView();
 }