+++ lib/src/network_manager.dart
@@ -2,27 +2,15 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
-// import 'package:dio/adapter.dart';
-// import 'package:dio/adapter_browser.dart';
-// import 'package:dio/adapter.dart';
-// import 'package:dio/dio.dart'
-// // ignore: uri_does_not_exist
-//     if (dart.library.html) 'package:dio/adapter_browser.dart'
-// // ignore: uri_does_not_exist
-//     if (dart.library.io) 'package:dio/adapter.dart';
-
 import 'package:dio/dio.dart';
-import 'package:dio/src/adapters/io_adapter.dart' if (dart.library.html) 'package:dio/src/adapters/browser_adapter.dart' as adapter;
+// ignore: implementation_imports
+import 'package:dio/src/adapters/io_adapter.dart'
+    if (dart.library.html) 'package:dio/src/adapters/browser_adapter.dart'
+    as adapter;
+// dart:io html, mobil
+//       pwa html css js, apk ipa
+// dart:html
 
-// import 'package:dio/src/adapter.dart'
-//     if (dart.library.io) 'package:dio/src/adapters/io_adapter.dart'
-//     if (dart.library.html) 'package:dio/src/adapters/browser_adapter.dart';
-
-import 'package:dio/src/dio_mixin.dart';
-import 'package:dio/src/dio.dart';
-import 'package:dio/src/dio_error.dart';
-import 'package:dio/src/interceptors/log.dart';
-import 'package:dio/src/cancel_token.dart';
 import 'package:flutter/foundation.dart';
 import 'package:logger/logger.dart';
 
@@ -51,7 +39,8 @@ class NetworkManager with DioMixin implements Dio, INetworkManager {
   /// [Future<DioError> Function(DioError error, NetworkManager newService)] of retry service request with new instance
   ///
   /// Default value function is null until to define your business.
-  late Future<DioError> Function(DioError error, NetworkManager newService)? onRefreshToken;
+  late Future<DioError> Function(DioError error, NetworkManager newService)?
+      onRefreshToken;
 
   /// [VoidCallback?] has send error if it has [onRefreshToken] callback after has problem.
   ///
@@ -102,7 +91,9 @@ class NetworkManager with DioMixin implements Dio, INetworkManager {
     _addLoggerInterceptor(isEnableLogger ?? false);
     _addNetworkInterceptors(interceptor);
 
-    httpClientAdapter = (!kIsWeb && customHttpClientAdapter != null) ? customHttpClientAdapter! : adapter.createAdapter();
+    httpClientAdapter = (!kIsWeb && customHttpClientAdapter != null)
+        ? customHttpClientAdapter!
+        : adapter.createAdapter();
   }
 
   void _addLoggerInterceptor(bool isEnableLogger) {
@@ -166,12 +157,15 @@ class NetworkManager with DioMixin implements Dio, INetworkManager {
     final body = _getBodyModel(data);
 
     try {
-      final response = await request('$path$urlSuffix', data: body, options: options, queryParameters: queryParameters);
-      if (response.statusCode! >= HttpStatus.ok && response.statusCode! <= HttpStatus.multipleChoices) {
+      final response = await request('$path$urlSuffix',
+          data: body, options: options, queryParameters: queryParameters);
+      if (response.statusCode! >= HttpStatus.ok &&
+          response.statusCode! <= HttpStatus.multipleChoices) {
         await writeCacheAll(expiration, response.data, method);
         return _getResponseResult<T, R>(response.data, parseModel);
       } else {
-        return ResponseModel(error: ErrorModel(description: response.data.toString()));
+        return ResponseModel(
+            error: ErrorModel(description: response.data.toString()));
       }
     } on DioError catch (e) {
       return _onError<R>(e);
@@ -179,13 +173,18 @@ class NetworkManager with DioMixin implements Dio, INetworkManager {
   }
 
   @override
-  Future<Response<Uint8List>> downloadFileSimple(String path, ProgressCallback? callback) async {
-    final response = await Dio().get<Uint8List>(path, options: Options(followRedirects: false, responseType: ResponseType.bytes), onReceiveProgress: callback);
+  Future<Response<Uint8List>> downloadFileSimple(
+      String path, ProgressCallback? callback) async {
+    final response = await Dio().get<Uint8List>(path,
+        options:
+            Options(followRedirects: false, responseType: ResponseType.bytes),
+        onReceiveProgress: callback);
 
     return response;
   }
 
-  Future<ResponseModel<R>?> _getCacheData<R, T extends INetworkModel>(Duration? expiration, RequestType type, T responseModel) async {
+  Future<ResponseModel<R>?> _getCacheData<R, T extends INetworkModel>(
+      Duration? expiration, RequestType type, T responseModel) async {
     // TODO: Web Cache support
     if (kIsWeb) return null;
     if (expiration == null) return null;
@@ -193,11 +192,13 @@ class NetworkManager with DioMixin implements Dio, INetworkManager {
     if (cacheDataString == null) {
       return null;
     } else {
-      return _getResponseResult<T, R>(jsonDecode(cacheDataString), responseModel);
+      return _getResponseResult<T, R>(
+          jsonDecode(cacheDataString), responseModel);
     }
   }
 
-  ResponseModel<R> _getResponseResult<T extends INetworkModel, R>(dynamic data, T parserModel) {
+  ResponseModel<R> _getResponseResult<T extends INetworkModel, R>(
+      dynamic data, T parserModel) {
     final model = _parseBody<R, T>(data, parserModel);
     return ResponseModel<R>(data: model);
   }
@@ -205,7 +206,11 @@ class NetworkManager with DioMixin implements Dio, INetworkManager {
   ResponseModel<R> _onError<R>(DioError e) {
     final errorResponse = e.response;
     _printErrorMessage(e.message);
-    final error = ErrorModel(description: e.message, statusCode: errorResponse != null ? errorResponse.statusCode : HttpStatus.internalServerError);
+    final error = ErrorModel(
+        description: e.message,
+        statusCode: errorResponse != null
+            ? errorResponse.statusCode
+            : HttpStatus.internalServerError);
     if (errorResponse != null) {
       _generateErrorModel(error, errorResponse.data);
     }