+++ lib/src/operation/network_model_parser.dart
@@ -12,17 +12,16 @@ extension _CoreServiceExtension on NetworkManager {
   }
 
   R? _parseBody<R, T extends INetworkModel>(dynamic responseBody, T model) {
-    if (responseBody is List) {
-      return responseBody.map((data) => model.fromJson(data)).cast<T>().toList() as R;
-    } else if (responseBody is Map<String, dynamic>) {
-      return model.fromJson(responseBody) as R;
-    } else {
-      try {
+    try {
+      if (responseBody is List) {
+        return responseBody.map((data) => model.fromJson(data)).cast<T>().toList() as R;
+      } else if (responseBody is Map<String, dynamic>) {
+        return model.fromJson(responseBody) as R;
+      } else {
         return EmptyModel(name: responseBody.toString()) as R;
-      } catch (e) {
-        print('Exception Vexana: $e');
-        return null;
       }
+    } catch (e) {
+      Logger().e('Parse Error: $e - response body: $responseBody T model: $T , R model: $R ');
     }
   }
 }